inputhook: move the new CTRL+C handling logic to this level
The inputhookqt4 module now only contains a simple, straightforward
inputhook_qt4 input hook doing only Qt4 event processing in short
batches (#481).

The logic to handle CTRL+C is now entirely done at the level of the
InputHookManager, where a new set_safe_inputhook() method can be used
to install a hook which will automatically benefit of the two-step
CTRL+C handling (first for interrupting the event loop and suspending
the input hook, second for returning to the prompt, at which point the
input hook is restored). This should help to address issues #122 and
catch KeyboardInterrupt and has to be installed using
set_safe_inputhook).

Some extra care is taken on Unix where we want to make sure no
KeyboardInterrupt is raised upon entering the Python callback (e.g. in
readline.c, if a KeyboardInterrupt is received during a select(), the
PyOS_InputHook is executed directly afterward and if it's a ctypes
Python callback the exception will be raised immediately on entering
the function, even before an initial `try:` could be executed).