inputhookqt4: improve responsiveness when pasting
The Qt4 inputhook now execs until the stdin is readable, instead of
waiting for a timeout to happen. This approach was supposed to play
better with readline which repeatedly calls the inputhook every 100ms.
But actually it does not. Same timings as before, roughly twice the
time it takes when there's no %gui. The timing for the original native
PyQt4 hook was somewhere in the middle (35ms - 58ms - 72ms).